Overwrite NPC arrival messages with custom ones
It seemed impossible to update an NPC's name in such way that
the arrival messages (e.g. "NAME the NPC has arrived!") would include
the modified name within. That's because the name gets modified in the
GlobalNPC.PreAI() overridden method, which is executed after.
The solution was to simply overwrite the already existing message
with one containing the custom NPC name.

Also removed one obsolete method from the ModContent class.