Add Copy, Paste and Carry buttons, fix NameField bug
Copy and Paste functionality is pretty straight-forward, I introduced
a new struct ListData which stores all the relevant data in a copy
and contains two methods for fetching and applying that data to the
actual in-game variables.
Carry mechanism exploits the fact that variables don't have to be reset
between saving and loading, so all it actually does is prevent
the usual reset from occurring. The code from CustomWorld.Initialize()
has been moved to CustomWorld ctor because turns out initialization
occurs everytime a world is loaded, which isn't intended and breaks
the carry functionality.

The NameField bug caused any new NameField panels to reset to empty
string every time the UINPCButton tab was refreshed. The reason was
the incorrect implementation of the "lastName" supplementary variable,
which has been fixed inside UINameField.Update() override.