Fix UINameList indexing after removing name entries
The Name List's order is determined by individual entries' index values
assigned at initialization. That value is always the size of
the currently existing list of names, and because of that the index
one-ups the previous highest index. However, I forgot that when erasing
entries from the list, the indexing should be shifted.