Fix name lists and randomization bugs, minor changes
* Changed default value of currentNames from null to "\0",
because null didn't save and load as a string properly,
* DeselectAllEntries() was causing a problem with invalid indexes
being added to currentNames. An additional check for SelectedNPC
has been added to RenamePanel.Deselect() to fix this,
* Simplified unique randomization code to fix some difficult crashes.
Surprisingly enough, the error messages were graphics-related. Still
don't know why, but they don't appear anymore so that's good,
* Overrode Mod.Unload() to free memory from many classes' objects,
* Removed X buttons's keybind text in parenthesis, because it caused
problems for users loading the mod for the first time,
* Added social_preview.png to the GitHub repository.

Overall a lot of stabilization patches.