Fix unbalanced packet queue, add receive chat message
CustomWorld.NetReceive has been properly tied with CustomWorld.NetSend,
because in the former implementation I made an error and would send too
many packets. Had to remove panelList.Print method from receive because
it cause troubles when starting up the server. Every receive call on
a client will now cause a "Receive ..." message to popup in the chat
for debugging purposes.

Also fixed a bug with trashicon where it would be misplaced at UIScales
different from 100%. A simple division was enough to bring it back to
order.