Add proper multiplayer syncing for mode and tryUnique
The reason why sync didn't work before was that I was merely requesting
the data on the server to be synchronized with the clients. Since world
data is always stored on the server, each request to sync would end up
in all the changes on clients being overwritten by server information,
i.e. reset completely. What I needed to do is send ModPackets of info
to the server, then request sync. To make that as easy and tidy as
possible, a new class PacketSender has been added to Network namespace.
Sending name dictionaries will be a bit tougher to devise, but mode
and tryUnique works brilliantly.
* Added HandlePacket overload to CustomNPCNames Mod class (this is
where the server receives packets sent to it),
* Changed code which used to request sync to code sending ModPackets,
* Add CustomNPCNames.Network.PacketSender class for packet management.