Make server the only shell to directly request sync
On closer inspection, some calls to SyncWorldData() wouldn't end
in a "Receiving..." message in the chat. The reason was that the caller
of the sync-initiating method was not affiliated with the server
(e.g. the ModPlayer class). To consistently prevent such silently
failing calls, a new PacketType, REQUEST_WORLD_SYNC was introduced.
From now on every execution of SyncWorldData() starts with a check
of NetmodeID to determine who the caller is. If it's the server,
the synchronization will commense as usual. If it's a client,
a REQUEST_WORLD_SYNC packet is sent to the server, which will in turn
call SyncWorldData() on its own.

Added some code (static ctors and one assignment operation)
to the SyncWorldData() method to make sure that any implicit execution
of ModWorld.NetSend() will be used with SyncType.Get == EVERYTHING.
This is necessary for the routine syncs that occur every minute or so.