PrintContent no longer interrupts editing names
There's always going to be multiplayer cases where a player is editing
some name while a server sync, or something else, triggers
an update of the UINameList, i.e. UINameList.PrintContent().
The consequences used to be deleting all entries and rebuilding
the entire list based on CustomWorld.CustomNames, but there's many
problems in that behavior, because newly added green entries don't
get added to CustomNames until they've been deselected, and all entries
suffer from getting deselected etc. This problem has been fixed in this
snapshot by storing the selected entry's data prior to a PrintContent()
and reapplying that data after the content has been printed anew.
The data is reapplied in the RecalculateChildren() hook, because
attempting to do so at the end of the PrintContent() method collided
with the logic implemented for RenameUI.panelListReady.

Other changes:
* Added a (currentName[id] != "\0") condition to the CustomNPC's
SetDefaults() method, because often the method would run unnecessarily,
causing already established NPC names to be randomized again, which
is absolutely unacceptable (this patch is still being tested).