Fix UINameField.lastName defaulting to null value
When using arrow keys to exit from a green entry, or sporadically
in other cases, sometimes a world sync would occur due to lastName
not matching the editName of an entry. Turned out the issue was caused
by entries getting recreated when PrintContent() executes, because
there was no default value explicitly set for lastName, and thus
it defaulted to null. I patched it by adding a simple assignment
in UINameField.ctor().