Replace currentNames with vanilla GivenName property
NPCs.CustomNPC.currentNames has been completely removed, because
I found a better way to handle name changing. With aid of IL editing
it's possible to directly swap the way vanilla names are generated
with a custom method. The IL editing is not implemented yet,
this snapshot is a necessary preperation step in the process of adding
it. Quite a lot of functionality has been removed along with
currentNames, because it was in some way connected to it.