Replace vanilla getNewNPCName with a custom hook
Turns out in spite of me having prepared perfectly well for IL editing,
an easier solution emerged in form of the On namespace. I effortlessly
replaced the default NPC.getNewNPCName method with my own one.

In the process of earlier IL editing preparations, I changed most
classes' visibilities to public. This is a redundant change, but it's
harmless, so here it will stay.

There's still a few bugs left to fix, namely removing list entries.