Fix multiple randomization functionality
In commit 94fa2e2 randomization was simplified, but unintentionally
broken. I patched it inside the RandomizeName method.

Also added a HasCustomNames method which checks for available custom
names with respect to the current mode and parameterized NPCID.

Replaced all occurrences of CountNPCs() with AnyNPCs(), because that's
the most effective way of checking if an NPC exists. I didn't use it
before, because I wasn't aware of its existence.