Fix first new entry not reappearing when printing
After adding the first, new entry in any name list and attempting
a reprint of the entire list, the new green entry would not reappear.
Furthermore, after adding a new entry or switching lists to some that
already contain entries, the entry would then make it back.
The reason behind this glitch was that panelListReady variable was only
ever set to false if the edited list had some entries saved to the
CustomNames dictionary, and since an entry doesn't save to that
dictionary until it's been deselected, for single-entry list that only
contain green members the CustomNames dictionary would be empty.

Simple fix: added a line setting panelListReady to false even when
there's no positions on the CustomNames dictionary's list.