Fix pasting in the same world for multiplayer clients
There's been a few issues with pasting everything on multiplayer.
After having done so, the IDs of all StringWrappers would be set to 0,
and thus all entries would become completely desynced from the server.
The reason syncing the incorrect server data back to the clients was
impossible is, constructing StringWrappers with 0 as ID causes it to
generate a new unique ID number instead of assigning 0.
I've actually discovered quite a few problems on my way to solving this
bug, here are all the things that have been altered in this snapshot:

* Changed the name of PacketType SEND_EVERYTHING to SEND_COPY_DATA.
This packet was used only for pasting anyway, and the critical error
I've made while writing the SendPacketToServer method was, it would
send current values instead of the values stored inside the copyData,
effectively making it useless. This has been fixed, naturally.
* New PacketType - SEND_COPY_NAMES. Counterpart for SEND_CUSTOM_NAMES
which was previously used by SEND_EVERYTHING to fix the error described
in point 1,
* ListData.customNames has been changed to a list of StringWrappers
as opposed to a list of strings for the obvious reason of storing
StringWrapper IDs of the copied objects,
* Removed UINPCButton.Refresh method, because all I needed it to ever
do could be done quicker by calling NameList.PrintContents,
* Added a Clone method to the StringWrapper class for easier deep
copying,
* PacketSender will now save StringWrapper.IDs when sending custom
names and copy names, as opposed to just strings. I can't believe I've
missed this error this whole time,
* CustomNPCNames.HandlePacket reads the now sent IDs from PacketSender
SEND_CUSTOM_NAMES and SEND_COPY_NAMES packets,
* Added a SendPacket 100 for debugging purposes, this is to be removed
in the final release.