Add busy functionality to all entry panels
1. THE BUSY SYSTEM
The busy functionality fixes the issue of having two players trying to
edit the same entry simultaneously. Now, every machine's CustomWorld
is equipped with a busyFields list, which contains IDs of either
StringWrappers which represent UINameFields that are being edited,
or just NPCIDs which represent UIRenamePanels that are being edited.

When any entry detects that its ID is on the busyFields list, it shuts
down all the editing functionality and flickers between its former
content and a "Player X is editing..." message, giving off a visual
 cue that someone else is doing something inside it. While in a busy
 state, each entry will stay constantly deselected for every client
 except the editor themselves, no matter what.

To make sure no delayed packets that potentially change the edited
entry's contents arrive after someone starts editing and ruin their
work, a lot of busy checks have been added to Mod.HandlePacket method,
and a few more Server->Client packets have been added (ones that start
with SERVER_REJECT_*). These packets get sent to only the client who
requested some operation that can't be executed, and display a simple
message informing the client that their request has been cancelled.

2. COPY SERVER TIMEOUT
A new mechanic has been added specifically for sending copy data.
After the initial SEND_COPY_DATA packet, which triggers the server's
response to inform the client when to start sending copy packets,
a hypothetical error is possible, where something interrupts the server
and the SERVER_AWAITING_COPY_DATA packet never arrives.
This is a problem, because the client's waitForServerResponse variable
effectively prevents them from doing anything until the server responds
which in this scenario never happens.

To fix this, a simple timer was added, _serverResponseTimeout, which
is set to 10 seconds after the SEND_COPY_DATA packets are sent.
In an event nothing happens and the timer reaches 0, the client's
waitForServerResponse variable gets toggled off, enabling the client
to continue working as normal. A message is also sent, informing
that client that either it's experiencing connection problems,
or there's a bug in the mod.