Configure busy mechanic to respond to foreign players
Up till now the busy mechanic was working, but for testing purposes
even the player who's editing considered himself an obstacle from
editing some entry. Now, I've added proper communication between client
and server, switching the busyFields list when appropriate so that the
actual entries being edited are being maintained there, as opposed to
manually selecting some to test results.

* Most conditions that required a "BusyField.ID == StringWrapper.ID"
or similar, have had "&& BusyField.player != Main.myPlayer" condition
added. This is so that players don't consider themselves as foreign
players,
* Modified the UIEntryPanel derived classes to send information to the
server when someone's Selecting or Deselecting them. A few adjustments
have been made to ensure that all possible ways of selecting or
deselecting an entry include an appropriate packet,
* Optimized the flow of RenameUI.UpdateUIStates() a tiny bit.

DISCLAIMER
This snapshot is actually impossible to troubleshoot solo, it's likely
that actual multiplayer testing will reveal some bugs which will be
dealt with in a future snapshot. This version is not confirmed to be
stable or working.