Fix and refactor busy searching
I made a mistake in the code checking if it's valid to clear a list
and if it's clear to cut/paste. The error was plain, I iterated through
wrong iterables when searching for an ID match in those places.
To make the code more readable and prevent such errors in the future,
I've added a few methods to CustomWorld which search the busyFields
list and return commonly needed values:
* AnyNameFieldBusy()
* AnyNameFieldBusyInCurrentTab()
* IsRenamePanelBusy()
* GetBusyField() - 2 overloads for player or ID parameters