Add missing conditions for locking randomization
When implementing some logic in RenameUI.UpdateUIStates I forgot to
include one more condition when determining whether or not there are
names too choose from (in Male, Female or Global tabs, i.e. group
randomization).
Here's what the condition used to be:
"If there exist Male/Female NPCs which have names on their custom list,
then there are names to choose from.".
Here's the fix:
"If there exist Male/Female NPCs which ARE ALIVE and have names on
their custom list, there are names to choose from.".

Yeah, I forgot the living aspect. I've restructured the code a bit
to make it faster too, but that's the only major tweak in logic.