Remove Clear All lock, replace it with smart removal
The idea is, instead of preventing a user from clearing a list while
someone else is editing, the user can still click Clear All and every
entry except the edited ones is going to get erased.

The same logic will not apply to things like Randomization, because
e.g. when randomizing from the Male tab, the only thing we know is
that some names are being edited, but the user doesn't know which,
so it seems unreliable to allow the user to randomize all male names
and gamble the results like that. And it would be tedious to implement.

+ Fixed one logically malformed for loop in CustomNPCNames line 417.