Update config files implementation to the native one
tModLoader 0.11.3 introduced extensive support for config files,
and it comes with a lot of simplicity and additional benefits, so
I happily switched over to it. There's no issues with server messaging
flags now, because server side values have been introduced.

There's some inconsistencies in the tModLoader's system still, so while
everything is pretty much working, some future patches will be needed
in order to bring the functionality to a perfect level (that's up to
the tModLoader staff and developers though).

The only inconvenience is that the GUI config values cannot be directly
mutated by outside code, e.g. changing carry's value via mod menu will
NOT update GUI config's Carry value. It syncs the other way around
though, so it's not a huge issue.