Start patching singleplayer, minor fixes
I had gotten so fixated on multiplayer that in the process of
implementing it I completely broke some singleplayer features (most of
these can probably be fixed with checking if the world is in multiplayer
mode or not). Anyway, the first major glitch that crashes only on
singleplayer was removed by removing a call to "DeselectAll()" inside
RenameUI.cs. The call was probably superfluous, so it's probably safe to
delete it. As to why it was causing an error, I'm not sure. The trace
message complained about trying to access a null pointer, which I can
understand, but it shouldn't be allowed, since all data is initialized
inside the "OnInitialize()" hook. Strange.

Made the "Deselect()" method in UINameList.cs actually pass the save
parameter further to the UINameField objects instead of not doing that
and letting them implicitly assume save=true. This is a silly mistake
that could've been there for ages without me noticing, but I just
noticed it now.

Added a little bit more informative comment in UINameList to describe
what the "savedSelectedNPC" variable does, because after coming back to
the code after 9 months I had difficulty gathering information on what
it does (mostly due to VisualStudio's incompetent search engine though).