Add netMode checks to busy packet sends and more
PacketSender cannot send packets in a singleplayer session, and it does
not silently fail on its own. My approach until now was to simply
precede any server communication with a Main.netMode check, but now I've
also added a check inside PacketSender's sending function that will
catch any mistakes or overlooks when locking that communication. It will
log such occurrences as warnings in the client log and silently exit the
function.